telegram-bot/requirements.txt
python-telegram-bot[socks]==13.14
emoji==2.1.0